Добавляет сброс слайдера
