+++ js/form-validate.js
@@ -1,8 +1,8 @@
 import { sendData } from './api.js';
 import {adForm} from './form.js';
 import { isEscapeKey } from './util.js';
-// import { sliderElement } from './slider.js';
 import { buttonReset } from './map.js';
+// import { sliderElement } from './slider.js';
 
 const pristine = new Pristine(adForm, {
   classTo: 'ad-form__element', // Элемент, на который будут добавляться классы
@@ -47,6 +47,36 @@ pristine.addValidator(price, validatePrice, getPriceErrorMessage);
 typeOfHousing.addEventListener('change', () => {
   price.placeholder = minPriceOfHousing[typeOfHousing.value];
 });
+
+const sliderElement = document.querySelector('.ad-form__slider');
+
+noUiSlider.create(sliderElement, {
+  range: {
+    min: 0,
+    max: 100000,
+  },
+  start: price.placeholder,
+  step: 10,
+  connect: 'lower',
+  format: {
+    to: function (value) {
+      return value.toFixed(0);
+    },
+    from: function (value) {
+      return parseFloat(value);
+    },
+  },
+});
+
+sliderElement.noUiSlider.on('update', () => {
+  price.value = sliderElement.noUiSlider.get();
+});
+
+price.addEventListener('change', () => sliderElement.noUiSlider.set(price.value));
+
+const resetSlider = () => {
+  sliderElement.noUiSlider.reset();
+};
 
 //Валидация полей количество комнат и количество мест
 const roomNumber = adForm.querySelector('#room_number');
@@ -98,8 +128,7 @@ const errorMessage = errorTemplate.cloneNode(true);
 buttonReset.addEventListener('click', (evt) => {
   evt.preventDefault();
   adForm.reset();
-  // resetMap();
-  // resetSlider();
+  resetSlider();
 });
 
 const onSuccessMessageClick = () => {
@@ -119,8 +148,7 @@ const sendFormSuccess = () => {
   document.addEventListener('click', onSuccessMessageClick);
   document.addEventListener('keydown', onSuccessMessageKeydown);
   adForm.reset();
-  // resetMap();
-  // resetSlider();
+  resetSlider();
 };
 
 const onErrorMessageClick = () => {
