+++ js/slider.js
@@ -1,28 +1,28 @@
-import { price } from './form-validate.js';
+// import { price } from './form-validate.js';
 
-//global noUiSlider:readonly
-export const sliderElement = document.querySelector('.ad-form__slider');
+// //global noUiSlider:readonly
+// export const sliderElement = document.querySelector('.ad-form__slider');
 
-noUiSlider.create(sliderElement, {
-  range: {
-    min: 0,
-    max: 100000,
-  },
-  start: price.placeholder,
-  step: 10,
-  connect: 'lower',
-  format: {
-    to: function (value) {
-      return value.toFixed(0);
-    },
-    from: function (value) {
-      return parseFloat(value);
-    },
-  },
-});
+// noUiSlider.create(sliderElement, {
+//   range: {
+//     min: 0,
+//     max: 100000,
+//   },
+//   start: price.placeholder,
+//   step: 10,
+//   connect: 'lower',
+//   format: {
+//     to: function (value) {
+//       return value.toFixed(0);
+//     },
+//     from: function (value) {
+//       return parseFloat(value);
+//     },
+//   },
+// });
 
-sliderElement.noUiSlider.on('update', () => {
-  price.value = sliderElement.noUiSlider.get();
-});
+// sliderElement.noUiSlider.on('update', () => {
+//   price.value = sliderElement.noUiSlider.get();
+// });
 
-price.addEventListener('change', () => sliderElement.noUiSlider.set(price.value));
+// price.addEventListener('change', () => sliderElement.noUiSlider.set(price.value));
